Refactor input fields
