+++ Client/src/components/Shared/TerminalBanner.tsx
